Adds dependency on Steven G. Harms' verbvector library.
A LatinVerb can be conceived as a state defined by four principal parts whose
activites are defined by ~121 methods or "vectors."  Writing out these method
names by hand is incredibly tedious, nevertheless, the ability to respond_to
when queried is very useful.

Further, these names can be used as proxies from larger clusters.  Verbvector
suggests clustering on tense.  Therefore, while a given vector may not, in the
code, respond but rather delegates to its "cluster" parent, it simulates as if
it does respond to all the vectors that one might logically deduce from the
grammatical structures of the Latin.