Expands test of FirstConj to test serialized form
In testFirstConjugation.rb, we run the exhaustive test of all the vectors.  We
then take the paradigumatic example defined in setup and we serialize it.  We
then deserialize that string and run _the very same_ tests on the revivified
string.  This is a good baseline for testing the TDD wrt. the serialization.