Converts single test file to Minitest
My test suite was failing because with the
ruby 1.9.2p180, and the new rdoc, we are now
using hoe and minitest.  As such, the test suite
of choice is Minitest.

Correcting the raises and other tests within
testLatinVerb.rb proved sufficient to get the
tests working again, but I belive I should open
up the other tests and convert to the standard.