Preliminary logic around deponent support
Deponent support is a difficult problem; however,
it can be handled in phases when you consider the
following key behaviors:

1.  Create passive forms for the "hidden" verb
within a deponent.  Given 'mirarī', the real verb
is 'mirāre.'  Therefore all you have to do is
derive the "real" from the deponent and then
re-assign the methods to point to the "real"
verb's passive forms.  That's sneaky.  That's
deponent_swap

2.  Some deponent verb forms are derived from
other results.  They're one-offs, but regular in
their aberrant behavior.  We stuff them into
deponent-mutations.  We eclipse the instance's
method by injecting methods by the same name into
the singleton class.  This happens in
deponent_mutations.

3.  Lastly, we have to tell initialize by string
that it should not attempt to derive the parts as
by a standard verb, but that it should create the
fake/real proxy (so that the mask has something to
reference).

This is the preliminary framing of the solution.