Handles ambiguous vector specification gracefully.
Previously you had to specify all 5 parts
(voice/mood/tense/person/number) as a method you
applied to a LatinVerb receiver.  While this is
pretty handy, there are certain "cluster" vectors
that are very informative as well.  To wit:
 * all the values in a tense
 * all the values in a tense for a given person
 * all the values in a tense for a given number

This commit adds the capability for these
ambiguous cases to yield their appropriate value.