LatinVerb#chart now returns nil
If one did puts @lv.chart, one would get the
return value of one of the #map calls from within
the method.  To prevent this, chart now returns
nil so that is can be puts without strange extra
visual weirdness.  It should be called without
puts though.

The unit test has been commented out because it
generates a lot of noise.  It may be handy for
future debugging so it's still there, though.