Add mutators/invariant.rb
I want to attach methods on to a LatinVerb in a
declarative fashion.  To do this we introduce the
`Mutators::Invariant` class.  It will use a series
of strategies to apply mutations onto the Verb.

Each of these strategies will implement a
`TenseBlock`.

Over a period of commits we will empty out the
present implementation of
`TenseDefinitions::Invariant` (an included module)
and move to the class (`Mutators::Invariant`)
method.