rm calls to set_as_defective
There was some weird action-at-a-distance going on
with the setting of the classifcation on the verb.
Now classifications are set *once* by the
classifier.

**THE ONLY EXCEPTION** is for irregular,
present-only verbs.  These are a huge, glaring
exception in the language.

Here, the classifier does as expected and returns
that it is an Irregular verb (true) and
consequently the mutator for Irregular is chosen
(correctly).  As part of the `mutate!` function,
the present-only-and-irregular verb has its
classification overridden to be PresentOnly.  The
`PresentOnlyIrregularMask` bears the responsbility
of performing that mutation.

Conflicts:

	lib/latinverb/tense_method_applicator.rb