MutatorForVerbType: do not use `extend` to add functionality
In the original implementation of `LatinVerb`, we
used modules to extend the base `LatinVerb` in order
to gain gain the various `Ten`seBlocks'
functionality.

In the case of Irregular verbs this wound up
looking super crazy because it required the use of
the callback hook `self.extended`.  It is far
preferable to explicitly add functionality (via
`instance_eval`) than rely on this hook method
"magic."