Remove InfinitiveBlock from infinitives' definitions
I would like to remove any non-vanilla structures
from the serialized infinitives.  I want to
extract the Linguistics component from the
LatinVerb component.

I want to have Latin, the language's, information
as a gem which can exist separate of
implementation of conjugating software.  This
Block thing is an impediment to that.