Extract "linguistics_latin" gem
LatinVerb is a client application.  It assumes
certain facts about the Latin language and the
verbs that populate it.  These facts do not belong
to `LatinVerb` itself, but rather to Latin::Verbs,
or, more correctly `Linguistics::Latin::Verb` or
`Linguistics::LA::Verb` (ISO-6391).

Thus we extract that gem in hopes that it will
line up with other `Linguistics` namespaced work
done by the Ruby community :)

E PLURIBUS LINGUIS MENTES MULTAS