Introduce Querent abstraction
`TenseMethodApplicator` generates a Querent
instance.  It is foreseeable that, in time, it
will be replaced and renamed "QuerentFactory" or
something similar as its purpose is to create the
querent instance (conveivably becoming `@querent`
on the `LatinVerb` instance.

That said, for the moment, `TenseMethod`
applicator resolves, by means of a
`QuerentFactory`, which type of Querent to make
based on the `VerbType`.  The querent is the thing
that does actual `TenseBlock` resolution via
`Forwardable`.

Since `VerbType`s inform the names of the
querents, the querents are the portmanteau of the
`VerbType` and the word "Querent."  We then mirror
the present design where the verb is thrown into
various classes and has its `TenseBlock` methods
pruned or altered.

Theq `Querent`s are all subclasses of a common
base which takes all the functionality that was
formerly in the "Invariant" strategy.

For the parameters specifying the person and
number of the `TenseBlock`, we implement
`QuerentTenseMethodsVectorizer`.