IrregularQuerent: remove weird tense declarations
It was really odd to have
`active_voice_indicative_mood_present_tense` and
`active_voice_indicative_mood_present_tense_third_person_singular_number`
defined there.

This brokenness was added to make sure that
semideponent tests worked.  Now that they are
removed it's much less gross.

It was also a good time to refactor the
SemideponentQuerentAdapter.  The use of "wrapping
methods" was less preferable to using delegation.

Impersonal Support
==================

Impersonal verbs are problematic in that they are
irregular, but that they also need to have their
active/indic./present tense-block bolted on so
that they honor the contract of `Querent`.  We
accomplish this by mixing in the impersonal
support if needed.