Implement Querent interface
1. LatinVerb: responds to #querent and delegates
that call to `@tense_method_applicator` which
supports `#querent`. Internally, it uses a
`QuerentFactory` which brokers the selection
of a `Querent` subclass (each _regular_ verb
type).  The returned TenseBlocks then are then
thrown to a "vectorizer" that adds the
permutations of person and number to look up
specific values in the returned `TenseBlock`.